Extract dependency check logic as Dependencies class
